Rework render system to have passes
General code cleanup
