+++ client/graphics/gbuffer.cpp
@@ -96,7 +96,7 @@ void GBuffer::StencilPass() {
 	glStencilOpSeparate(GL_FRONT, GL_KEEP, GL_DECR_WRAP, GL_KEEP);
 }
 
-void GBuffer::BeginLightPass() {
+void GBuffer::BeginLightPass() const {
 	glDrawBuffer(GL_COLOR_ATTACHMENT4);
 	glDisable(GL_DEPTH_TEST);
 	glEnable(GL_CULL_FACE);
@@ -111,11 +111,11 @@ void GBuffer::BeginLightPass() {
 	}
 }
 
-void GBuffer::BeginDirLightPass() { glStencilFunc(GL_ALWAYS, 0, 0); }
+void GBuffer::BeginDirLightPass() const { glStencilFunc(GL_ALWAYS, 0, 0); }
 
-void GBuffer::BeginPointLightPass() {}
+void GBuffer::BeginPointLightPass() const {}
 
-void GBuffer::EndPointLightPass() {
+void GBuffer::EndPointLightPass() const {
 	// glDisable(GL_BLEND);
 }
