+++ client/graphics/view.hpp
@@ -6,9 +6,9 @@
 
 namespace tec {
 struct View {
-	View(bool active = false) : active(active) {}
+	explicit View(const bool active = false) : active(active) {}
 	glm::vec3 view_pos{0};
-	glm::quat view_quat;
+	glm::quat view_quat{};
 	bool active = false;
 
 	void Out(proto::Component* target) {
